check if resource exist before deleting it

+++ pkg/reconciler/kubernetes/adapter.go
@@ -214,6 +214,9 @@ func (g *kubeClientAdapter) newNamespaceUnstruct(namespace string) (*unstructure
 }
 
 func (g *kubeClientAdapter) DeleteResource(kind, name, namespace string) (*Resource, error) {
+	if !g.resourceExists(kind, name, namespace) {
+		return nil, nil
+	}
 	metadata, err := g.kubeClient.DeleteResourceByKindAndNameAndNamespace(kind, name, namespace, metav1.DeleteOptions{})
 	deletedResource := toResource(metadata)
 	if err != nil && !k8serr.IsNotFound(err) {
@@ -247,8 +250,7 @@ func (g *kubeClientAdapter) Delete(ctx context.Context, manifest, namespace stri
 		unstruct := unstructs[i]
 
 		// execute the delete request only if the resource exists
-		_, err := g.kubeClient.Get(unstruct.GetKind(), unstruct.GetName(), namespace)
-		if k8serr.IsNotFound(err) {
+		if !g.resourceExists(unstruct.GetKind(), unstruct.GetName(), namespace) {
 			continue
 		}
 
@@ -287,6 +289,16 @@ func (g *kubeClientAdapter) Delete(ctx context.Context, manifest, namespace stri
 	}
 	return deletedResources, nil
 }
+
+// check if resource exists in the cluster
+func (g *kubeClientAdapter) resourceExists(kind, name, namespace string) bool {
+	_, err := g.kubeClient.Get(kind, name, namespace)
+	if k8serr.IsNotFound(err) {
+		return false
+	}
+	return err == nil
+}
+
 func (g *kubeClientAdapter) newProgressTracker() (*progress.Tracker, error) {
 	clientSet, err := g.Clientset()
 	if err != nil {
